fix: ScanStatusResponse done field as float; resolve episode paths for all series
- schemas: done field changed from int to float to accept 0.5 increments (was causing
  Pydantic ValidationError crashing /scan/status)
- pipeline Phase 1a: always resolve episode paths for all Series regardless of whether
  the series folder exists locally — fixes series whose mount path changed (e.g. Fallout
  showing no_file because scopuli folder gone, but Jellyfin knows episodes are at xtor)
- pipeline Phase 1b: use resolved episode path for Series even when series folder is gone;
  item_root falls back to episode parent dir in that case

+++ app/web/schemas.py
@@ -7,7 +7,7 @@ class ScanStatusResponse(BaseModel):
     running: bool
     cancelled: bool = False
     total: int
-    done: int
+    done: float
     current_item: str
     error: str | None = None
